requirements.txt
accelerate
bert_score
bitsandbytes
cython
einops
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
evaluate
ipykernel
matplotlib
nltk
peft
pillow
protobuf
pycocotools
qwen-vl-utils
scikit-learn
sentencepiece
scikit-image
spacy
torch
torchvision
transformers == 4.46.3
